+++ src/auth/strategy/jwt-google.strategy.ts
@@ -29,7 +29,7 @@ export class JwtGoogleStrategy extends PassportStrategy(Strategy, 'google') {
     const user = {
       // id: profile.id,
       email: emails[0].value,
-      nickname: name.familyName + name.givenName,
+      nickname: 'Google_User' + Math.random().toString(36).substring(2, 12),
       // nickName: name.familyName + name.givenName,
       //패스워드는 36진수의 랜덤숫자로 생성되어 저장할 예정
       password: Math.random().toString(36).substring(2, 12),
