--- src/user/dto/update-user.dto.ts
@@ -1,4 +1,0 @@
-
-export class UpdateUserDto {
-
-}
