+++ src/user/user.controller.ts
@@ -15,20 +15,5 @@ export class UserController {
     private readonly userService: UserService,
   ) {}
 
-  @Get('/me')
-  @Render('mypage.ejs')
-  @UseGuards(JwtAuthGuard)
-  async getInformation(@Cookies('Authentication') jwt: JwtDecodeDto) {
-    console.log('아디쓰', jwt.id);
-    const userId = jwt.id;
-    const User = await this.userService.getUserInformation(userId);
-    const data = {
-      nickname: User.nickname,
-      email: User.email,
-      phone: User.phone,
-      address: User.address,
-    };
-    console.log('결과적으론 데이터가 리턴중', data);
-    return { data: data };
-  }
+
 }
