+++ src/views/main-footer.ejs
@@ -30,7 +30,7 @@
                 justify-content: center;font-size: 35px; padding: 0">
                 <a onclick='deal()'>🚀</a>
             </div>
-            <div class="col" onclick="location.href='/mypage'"
+            <div class="col" onclick="location.href='/me'"
                  style="
                     border: 1px solid black; height: 50px; width: 25%;
                     display: flex;
