+++ src/views/signup.ejs
@@ -23,6 +23,7 @@
           .replace(/^(\d{0,3})(\d{0,4})(\d{0,4})$/g, "$1-$2-$3").replace(/(\-{1,2})$/g, "");
       }</script>
     <!--    Js -->
+
     <script><%- include('js/order-layout.js') %></script>
     <script><%- include('js/main-header.js') %></script>
     <title>Document</title>
